refactor: client follow logic
